Setting the app to its initial structure
Due to issues with setting the app up on server, the previous structure has been set
