--- flask_application/.idea/inspectionProfiles/Project_Default.xml
@@ -1,14 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredPackages">
-        <value>
-          <list size="1">
-            <item index="0" class="java.lang.String" itemvalue="python_bcrypt" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
