--- flask_application/run.py
@@ -1,5 +1,0 @@
-from app import app
-
-if __name__ == "__main__":
-    app.run()
-
