+++ flask_application/static/css/style.css
