+++ flask_application/static/img/restaurant.jpg
