+++ flask_application/static/js/favourites.js
