+++ flask_application/static/js/filter.js
