+++ flask_application/static/js/filterOpen.js
