+++ flask_application/static/js/filters.js
