+++ flask_application/static/js/logout.js
