+++ flask_application/static/js/redirect.js
