+++ flask_application/static/js/restCard.js
