+++ flask_application/static/js/signup.js
