+++ flask_application/templates/redirect.html
